Updated UI and Created logic for New notifications
There is now conditional logic that will alter visibility of notification options on the home screen, as well as logic that pushes incoming notifications from the main activity to TwitterNotifications. Since the visibility of views on the main activity is dependent on the existence of notifications, there needs to be tracking of them from the main activity.